+++ src/Geopilot.Pipeline/Processes/Matcher/XtfMatcher/XtfMatcherProcess.cs
@@ -53,7 +53,7 @@ internal class XtfMatcherProcess
     }
 
     [PipelineProcessRun]
-    public Task<Dictionary<string, object?>> RunAsync([UploadFiles] IPipelineFileList uploadFiles)
+    public Task<Dictionary<string, object?>> RunAsync(IPipelineFileList uploadFiles)
     {
         var filtered = uploadFiles;